Rename Tuya motion quirk (#3789)
