+++ zhaquirks/tuya/tuya_motion.py
